requirements.txt
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
peft>=0.7.0
datasets>=2.14.0
pillow>=10.0.0
bitsandbytes>=0.41.0
safetensors>=0.4.0
huggingface_hub>=0.19.0
wandb>=0.15.0
tqdm>=4.66.0
